LPS-139047 Check if table exists

+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/src/main/java/com/liferay/dynamic/data/mapping/internal/upgrade/v5_0_0/PollsToDDMUpgradeProcess.java
@@ -128,6 +128,10 @@ public class PollsToDDMUpgradeProcess extends UpgradeProcess {
 	protected void doUpgrade() throws Exception {
 		_upgradeDDMFormInstance();
 
+		if (!hasTable("PollsQuestion")) {
+			return;
+		}
+
 		_upgradePollsQuestions();
 
 		runSQL("delete from PollsVote");
